requirements.txt
django_countries
django-money
Pillow